Move the non-tests from component to component.config
