+++ airbyte-integrations/connectors/destination-snowflake/src/test-integration/kotlin/io/airbyte/integrations/destination/snowflake/component/config/SnowflakeTestConfigFactory.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2025 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.destination.snowflake.component
+package io.airbyte.integrations.destination.snowflake.component.config
 
 import io.airbyte.cdk.load.component.config.TestConfigLoader.loadTestConfig
 import io.airbyte.integrations.destination.snowflake.spec.SnowflakeConfiguration
